requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
faiss-cpu
streamlit
pydantic
python-dotenv
beautifulsoup4
requests